wip - grouped table files in folder, trying to fix ts error
